Add files via upload
This project is a deep learning-based Cat and Dog Image Classifier using Convolutional Neural Networks (CNNs) with TensorFlow & Keras. It classifies images into cat or dog categories by training on a labeled dataset. The model is optimized with data augmentation to improve accuracy and generalization. The dataset includes training, validation, and test sets, and the model achieves an accuracy of 63-70%. The project also visualizes accuracy and loss metrics to evaluate performance. 🚀